refactor: service, provider 역할 분리
